Move test code to VMTest class
