+++ rskj-core/src/main/java/org/ethereum/vm/program/Program.java
@@ -26,6 +26,7 @@ import co.rsk.peg.Bridge;
 import co.rsk.remasc.RemascContract;
 import co.rsk.vm.BitSet;
 import com.google.common.annotations.VisibleForTesting;
+import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.config.BlockchainConfig;
 import org.ethereum.config.Constants;
 import org.ethereum.core.Block;
@@ -35,7 +36,6 @@ import org.ethereum.crypto.HashUtil;
 import org.ethereum.db.ContractDetails;
 import org.ethereum.util.ByteUtil;
 import org.ethereum.util.FastByteComparisons;
-import org.ethereum.util.Utils;
 import org.ethereum.vm.*;
 import org.ethereum.vm.MessageCall.MsgType;
 import org.ethereum.vm.PrecompiledContracts.PrecompiledContract;
@@ -48,13 +48,10 @@ import org.ethereum.vm.trace.ProgramTrace;
 import org.ethereum.vm.trace.ProgramTraceListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.bouncycastle.util.encoders.Hex;
 
-import java.io.ByteArrayOutputStream;
 import java.math.BigInteger;
 import java.util.*;
 
-import static java.lang.StrictMath.min;
 import static java.lang.String.format;
 import static org.apache.commons.lang3.ArrayUtils.*;
 import static org.ethereum.util.BIUtil.*;
@@ -1341,76 +1338,6 @@ public class Program {
         }
     }
 
-    static String formatBinData(byte[] binData, int startPC) {
-        StringBuilder ret = new StringBuilder();
-        for (int i = 0; i < binData.length; i+= 16) {
-            ret.append(Utils.align("" + Integer.toHexString(startPC + (i)) + ":", ' ', 8, false));
-            ret.append(Hex.toHexString(binData, i, min(16, binData.length - i))).append('\n');
-        }
-        return ret.toString();
-    }
-
-    public static String stringifyMultiline(byte[] code) {
-        int index = 0;
-        StringBuilder sb = new StringBuilder();
-        BitSet mask = buildReachableBytecodesMask(code);
-        ByteArrayOutputStream binData = new ByteArrayOutputStream();
-        int binDataStartPC = -1;
-
-        while (index < code.length) {
-            final byte opCode = code[index];
-            OpCode op = OpCode.code(opCode);
-
-            if (!mask.get(index)) {
-                if (binDataStartPC == -1) {
-                    binDataStartPC = index;
-                }
-                binData.write(code[index]);
-                index ++;
-                if (index < code.length) {
-                    continue;
-                }
-            }
-
-            if (binDataStartPC != -1) {
-                sb.append(formatBinData(binData.toByteArray(), binDataStartPC));
-                binDataStartPC = -1;
-                binData = new ByteArrayOutputStream();
-                if (index == code.length) {
-                    continue;
-                }
-            }
-
-            sb.append(Utils.align("" + Integer.toHexString(index) + ":", ' ', 8, false));
-
-            if (op == null) {
-                sb.append("<UNKNOWN>: ").append(0xFF & opCode).append("\n");
-                index ++;
-                continue;
-            }
-
-            if (op.name().startsWith("PUSH")) {
-                sb.append(' ').append(op.name()).append(' ');
-
-                int nPush = op.val() - OpCode.PUSH1.val() + 1;
-                byte[] data = Arrays.copyOfRange(code, index + 1, index + nPush + 1);
-                BigInteger bi = new BigInteger(1, data);
-                sb.append("0x").append(bi.toString(16));
-                if (bi.bitLength() <= 32) {
-                    sb.append(" (").append(new BigInteger(1, data).toString()).append(") ");
-                }
-
-                index += nPush + 1;
-            } else {
-                sb.append(' ').append(op.name());
-                index++;
-            }
-            sb.append('\n');
-        }
-
-        return sb.toString();
-    }
-
     public DataWord getReturnDataBufferSize() {
         return new DataWord(getReturnDataBufferSizeI());
     }
@@ -1435,81 +1362,6 @@ public class Program {
 
     public BlockchainConfig getBlockchainConfig() {
         return blockchainConfig;
-    }
-
-    static class ByteCodeIterator {
-        private byte[] code;
-        private int pc;
-
-        public ByteCodeIterator(byte[] code) {
-            this.code = code;
-        }
-
-        public void setPC(int pc) {
-            this.pc = pc;
-        }
-
-        public int getPC() {
-            return pc;
-        }
-
-        public OpCode getCurOpcode() {
-            return pc < code.length ? OpCode.code(code[pc]) : null;
-        }
-
-        public boolean isPush() {
-            return getCurOpcode() != null ? getCurOpcode().name().startsWith("PUSH") : false;
-        }
-
-        public byte[] getCurOpcodeArg() {
-            if (isPush()) {
-                int nPush = getCurOpcode().val() - OpCode.PUSH1.val() + 1;
-                return Arrays.copyOfRange(code, pc + 1, pc + nPush + 1);
-            } else {
-                return new byte[0];
-            }
-        }
-
-        public boolean next() {
-            pc += 1 + getCurOpcodeArg().length;
-            return pc < code.length;
-        }
-    }
-
-    static BitSet buildReachableBytecodesMask(byte[] code) {
-        NavigableSet<Integer> gotos = new TreeSet<>();
-        ByteCodeIterator it = new ByteCodeIterator(code);
-        BitSet ret = new BitSet(code.length);
-        int lastPush = 0;
-        int lastPushPC = 0;
-        do {
-            ret.set(it.getPC()); // reachable bytecode
-            if (it.isPush()) {
-                lastPush = new BigInteger(1, it.getCurOpcodeArg()).intValue();
-                lastPushPC = it.getPC();
-            }
-            if (it.getCurOpcode() == OpCode.JUMP || it.getCurOpcode() == OpCode.JUMPI) {
-                if (it.getPC() != lastPushPC + 1) {
-                    // some PC arithmetic we totally can't deal with
-                    // assuming all bytecodes are reachable as a fallback
-                    ret.setAll();
-                    return ret;
-                }
-                int jumpPC = lastPush;
-                if (!ret.get(jumpPC)) {
-                    // code was not explored yet
-                    gotos.add(jumpPC);
-                }
-            }
-            if (it.getCurOpcode() == OpCode.JUMP || it.getCurOpcode() == OpCode.RETURN ||
-                    it.getCurOpcode() == OpCode.STOP) {
-                if (gotos.isEmpty()) {
-                    break;
-                }
-                it.setPC(gotos.pollFirst());
-            }
-        } while(it.next());
-        return ret;
     }
 
     public void addListener(ProgramOutListener listener) {
